+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## v1.8.0rc1
 
 * Adds a generic reader used when `metgenc` processes granules with an unknown
-  file-type. It uses spatial metadata from `spatial` files or collection metadata.
+  file-type. It uses spatial metadata from `spatial` files or collection metadata. (Issue-199)
 * Add `.ini` option to use collection temporal extent information to describe each granule. (Issue-139)
 
 ## v1.7.0